pyproject.toml
[project]
name = "py-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "django>=5.2.5",
    "google-genai>=1.30.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
